apply interceptors to all client
